Removed condition for Symbol Floating page to show only during trading hours
With the inclusion of Crypto, we'll need to show the crypto floating position over the weekends as well. so, I've removed the condition so that the details will show on the symbol floating page over the weekend. However, on javascript side, I've set the condition that if it's the weekend (Did't do GMT-UTC translation) the refresh rate would be 300 seconds. This is done to save SQL load over the weekend